% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simetrize_matrix.R
\name{simetrize_matrix}
\alias{simetrize_matrix}
\title{Get the full symetric matrix from a upper triangular contact matrix}
\usage{
simetrize_matrix(mat)
}
\arguments{
\item{mat}{HiC contact matrix (should be upper triangular)}
}
\value{
The full symetric HiC contact matrix corresponding to the input one
}
\description{
This function takes a upper triangular HiC contact matrix and returns the full symetric matrix.
}
\examples{
plot(0)
}

